+++ fsm/byzantine_test.go
@@ -2,6 +2,7 @@ package fsm
 
 import (
 	"bytes"
+	"fmt"
 	"github.com/canopy-network/canopy/fsm/types"
 	"github.com/canopy-network/canopy/lib"
 	"github.com/canopy-network/canopy/lib/crypto"
@@ -9,6 +10,10 @@ import (
 	"slices"
 	"testing"
 )
+
+func Test(t *testing.T) {
+	fmt.Println(lib.Uint64ReducePercentage(1, 1))
+}
 
 func TestHandleByzantine(t *testing.T) {
 	// IMPORTANT NOTE: the amount of case testing is limited here due to the amount of code covered in this function